+++ src/contexts/beatFilmsContext.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const beatFilmContext = React.createContext();
+
+export default beatFilmContext;
